feat(seed phrases): organize `SeedPhrase` into module
