+++ crypto/src/keys.rs
@@ -5,9 +5,10 @@ mod nullifier;
 pub use nullifier::{NullifierKey, NK_LEN_BYTES};
 
 mod seed_phrase;
+pub use seed_phrase::SeedPhrase;
 
 mod spend;
-pub use spend::{SeedPhrase, SpendKey, SpendSeed, SPENDSEED_LEN_BYTES};
+pub use spend::{SpendKey, SpendSeed, SPENDSEED_LEN_BYTES};
 
 mod fvk;
 mod ivk;
